Wad: support loading WAD2 (Quake) texture wads.
WAD2 is the same as WAD3 except that the miptex entries don't have a
palette after the pixel data, the palette needs to be supplied externally.

My implementation is lame / hacky, in that WadImageStream just looks for
the quake palette in {exe}/Palettes/quake.lmp. I included the palette with
this commit, it's supposed to be public domain
(see http://quakewiki.org/wiki/Quake_palette#Legal_status)

It would be cleaner to read from the user's quake directory, and this
would also allow displaying textures for a Quake TC that uses its
own palette.